requirements.txt
Django>=3.0.3
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
djangorestframework-gis==0.15
django-phonenumber-field==3.0.1
phonenumbers==8.11.0
psycopg2-binary==2.8.4
django-reversion==3.0.5
requests>=2.23.0
